% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2script.R
\name{package2script}
\alias{package2script}
\title{Produce R script with package source code}
\usage{
package2script(package, file = NULL, source = "local",
  replace = FALSE, directory = "..")
}
\arguments{
\item{package}{Name of the package.}

\item{file}{Path for the output file. Must be of type R-script (.R). Default name is
the name of the package.}

\item{source}{Where to find the package? As of now, only "local" is implemented,
corresponding to the package source code being available on the local machine.}

\item{replace}{Should an already existing version of the produced R script be replaced?
Defaults to \code{FALSE} (no).}

\item{directory}{Path to the directory in which the source code for the package in located.
Defaults to the parent directory of the current directory.}
}
\description{
Produce R script with package source code
}
